+++ inst/ubinc/templates/r_nlmixr.R
@@ -1,0 +1,9 @@
+#
+#
+#
+my_model <- function() {
+ini({<SYSTEM_PARAMS_TV><IIV_DEF><IIV_BLOCK><ERROR_PARAMS>
+})
+model({ <SYSTEM_PARAMS><SSP><DSP><ODES><OUTPUTS>
+})
+}